+++ NEWS.md
@@ -1,6 +1,7 @@
 # ps (development version)
 
-* New function `ps_disk_io_counters()` to query disk I/O counters implemented on Windows and Linux. (#145, @michaelwalshe).
+* New function `ps_disk_io_counters()` to query disk I/O counters
+  (#145, @michaelwalshe).
 
 * New `ps_fs_info()` to query information about the file system of one
   or more files or directories.